chatbot/requirements.txt
flask
requests
langchain
langchain-community
langchain-core
ollama
python-dotenv  # ✅ Load .env variables
